Fix netcdf-4.2.20.jar to include unidataCommon.
This is necessary because the ucar.unidata.io.RandomAccessFile class is
needed for successful compilation, and it was moved to unidataCommon.

As a side note, it seems to have been moved back into the core netcdf
JAR as of the latest version (4.2.27), so hopefully we will not need to
manually adjust the JAR contents for any future upgrades.

So, we could have added the unidataCommon JAR as a temporary, additional
dependency, but: 1) it is purely transitive (we do not reference any of
those classes directly in our code AFAIK); and 2) the JAR seems to have
disappeared as of 4.2.27, so we would need to remove it later anyway.