Fix variable names to work with netCDF 4.2.20.
In netCDF 4.0.41, there was a bug (I think) that caused certain
characters in variable names to be replaced with an underscore.  For
example, if the variable name in the file was "Image data #0", then the
variable name returned by netCDF was "Image_data_0".  The Imaris HDF
reader and the netCDF unit tests both coded around this problem, so it
was not very noticeable.

Now that we have upgraded to 4.2.20, the bug has seemingly been fixed and it
is necessary to fix the expected variable names accordingly.