WIP: strip namespaces after every transformation
I don't know whether this is really warranted. However, doing this
avoids a nasty problem with data from WiscScan, whereby 2012 schema XML
transforms to 2013, and then 2015, but then fails to transform to 2016.

Further effort is needed to create a minimal XML block + Java code that
makes the issue easily reproducible.